requirements.txt
selenium
requests
webdriver-manager
playwright
pyyaml
